audio feature: fix timer handle leaks.
we should close the timer handle and modify the timing of
releasing the audio feature object to avoid use after free,
because the Audio object is assigned to the timer's data
when the app is created(system_audio_onCreate) for use
in the callback. To avoid use-after-free, the object can
only be released when the timer is completely closed ,
that is, after the timer_close_cb callback is executed.

Signed-off-by: cenzhanquan1 <cenzhanquan1@xiaomi.com>