audio/feature:Delete the error event callback
Since the state has been changed and an error callback has been sent to the application when the open operation fails, the application will execute the stop operation. During the stop process, the changed state will be checked again, triggering another error callback to the application. This creates an infinite loop that leads to a crash.

Signed-off-by: zhangsenlin <zhangsenlin@xiaomi.com>